Finished adding pundit and all the policies

+++ app/services/base_destroyer.rb
@@ -1,34 +1,17 @@
-class BaseDestroyer
-  include CommonHelper
-  include PrerequisitesHelper
+  class BaseDestroyer < CommonBase
 
   attr_reader :record
 
-  def self.call
-    obj = self.class.new
-    obj.needed_before_call
-    obj.call
-    obj.needed_after_call
-  end
-
   def call(attrs)
     @record = model.find(attrs)
-    record = @record.destroy
-    success = record.present? && record.errors.blank?
-    unless success
-      add_error(message: "Failed to destroy #{self.class.name}!", traceback: @record.errors)
+    if @record.present?
+      record = @record.destroy
+      success = record.present? && record.errors.blank?
+      unless success
+        add_error(message: "Failed to destroy #{self.class.name}!", traceback: @record.errors)
+      end
+    else
+      add_error(message: "#{model.name} not found!", traceback: [])  
     end
-  end
   
-  def model
-    raise "Must be implemented in inheriting class"
-  end
-
-  def needed_before_call
-    prerequisites
-  end
-  
-  def needed_after_call
-  end
-
 end
